+++ contracts/adapters/paraswap/interfaces/IParaSwapAugustus.sol
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: agpl-3.0
+pragma solidity 0.8.10;
+
+interface IParaSwapAugustus {
+  function getTokenTransferProxy() external view returns (address);
+}