+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+### [1.12.1](https://www.github.com/aave/aave-v3-periphery/compare/v1.12.0...v1.12.1) (2022-02-25)
+
+
+### Bug Fixes
+
+* Rename lendingPool to pool in UiProvider contract ([2e50d3f](https://www.github.com/aave/aave-v3-periphery/commit/2e50d3f57bd0ccc4d2649ec8c7b36770158829ee))
+
+## [1.12.0](https://www.github.com/aave/aave-v3-periphery/compare/v1.11.1...v1.12.0) (2022-02-24)
+
+
+### Features
+
+* add Collector implementation and controller ([56a39c2](https://www.github.com/aave/aave-v3-periphery/commit/56a39c2daf502ffe3da36072ffdfd8e73d8b47b9))
+
 ### [1.11.1](https://www.github.com/aave/aave-v3-periphery/compare/v1.11.0...v1.11.1) (2022-02-15)